+++ CHANGES.md
@@ -2,7 +2,11 @@
 
 ## [Unreleased]
 
+### Changed
+
 - Disable transaction and bulk_transactions extensions by default **breaking change**
+- update `stac-fastapi-*` version requirements to `>=5.2,<6.0`
+- add pgstac health-check in `/_mgmt/health`
 
 ## [5.0.2] - 2025-04-07